Reduce time.sleep 's in tests
We had a few unnecessary time.sleep 's in tests, removed those to speed
up tests. The remaining ones are in listen integration tests and cannot
be removed entirely. Those sleeps are there to wait for timescale writer
to pick up and insert the listens. Reducing the sleep duration there a
bit to poll more frequently.