Relax ratelimit during tests
There are a few pinned recording api tests that fails occasionally due
to ratelimiting. This has also happened in the past with other tests
so relaxing ratelimiting during tests by treating all requests as if
from whitelisted users.

Note that the disadvantage of this against ratelimiting completely in
tests is that only requests with an Authorization header enjoy the
relaxed limit.